Modified for Trinket, start-up sequence, & enhanced flicker
Adafruit's "SuperHero Power Plant" code with a few customizations.
Changed some of color values to make a little brighter and whiter than
Adafruit's original code.
Added a start-up sequence so that it turns on bright white then fades
to blue.
Also, added an if statement to increase flicker/"stutter" effect.
Runs on Trinket.